routing when count is 0

+++ program/src/app/onboarding/viewcart/viewcart.page.ts
@@ -444,19 +444,24 @@ export class ViewcartPage implements OnInit {
   }
 
   remove(){
-    for(var i=0;i<this.cartList.length;i++){
-      if(this.cartList[i].MySelf=="True")
-      {
-        var id=this.cartList[i].CartId;
-        this.cartList[i].Qty==1
-        this.cartList.splice(i,1)
-        this.service.deleteItem({"Id":parseFloat(id)})
-        .subscribe(res=> {
-          this.service.isActivationFlow=true;
-          this.router.navigate(['/onboarding/add-to-cart']);
-        })
+    if(this.cartList.length==0){
+      this.router.navigate(['/onboarding/add-to-cart']);
+    }else{
+      for(var i=0;i<this.cartList.length;i++){
+        if(this.cartList[i].MySelf=="True")
+        {
+          var id=this.cartList[i].CartId;
+          this.cartList[i].Qty==1
+          this.cartList.splice(i,1)
+          this.service.deleteItem({"Id":parseFloat(id)})
+          .subscribe(res=> {
+            this.service.isActivationFlow=true;
+            this.router.navigate(['/onboarding/add-to-cart']);
+          })
+        }
       }
+      console.log(this.cartList)
     }
-    console.log(this.cartList)
+   
   }
 }
